+++ index.browser.ts
@@ -5,6 +5,7 @@ import {recoverJSON} from "./inc/jsonRecovery";
 import * as lilyParser from "./inc/lilyParser";
 import LogRecorder from "./inc/logRecorder";
 import * as constants from "./inc/constants";
+import StreamParser from "./inc/streamParser";
 
 import SheetLive from "./app/components/sheet-live.vue";
 import SheetSigns from "./app/components/sheet-signs.vue";
@@ -20,6 +21,7 @@ export {
 	lilyParser,
 	LogRecorder,
 	constants,
+	StreamParser,
 	SheetLive,
 	SheetSigns,
 	SheetBaker,